+++ tests/test_main_flow.py
@@ -3,43 +3,8 @@ import pytest
 import sys
 from unittest.mock import patch
 
-from tests.mocks.dummy_ai import DummyFlashcardAI
-from tests.mocks.dummy_obsidian import DummyObsidianAPI
-from tests.mocks.dummy_anki import DummyAnkiAPI
-
-
-@pytest.fixture
-def mock_services(monkeypatch):
-    """Set up mock services before each test"""
-    # Set dummy env vars so real API clients can be imported
-    monkeypatch.setenv("OBSIDIAN_API_KEY", "test_key")
-    monkeypatch.setenv("ANTHROPIC_API_KEY", "test_key")
-
-    import obsidianki.cli.services
-    import sys
-    import importlib
-
-    original_ai = obsidianki.cli.services.AI
-    original_obsidian = obsidianki.cli.services.OBSIDIAN
-    original_anki = obsidianki.cli.services.ANKI
-
-    # Create fresh instances for each test
-    obsidianki.cli.services.AI = DummyFlashcardAI()
-    obsidianki.cli.services.OBSIDIAN = DummyObsidianAPI()
-    obsidianki.cli.services.ANKI = DummyAnkiAPI()
-
-    # Force reload of modules that import obsidianki.cli.services to pick up fresh instances
-    if 'obsidianki.cli.processors' in sys.modules:
-        importlib.reload(sys.modules['obsidianki.cli.processors'])
-    if 'obsidianki.main' in sys.modules:
-        importlib.reload(sys.modules['obsidianki.main'])
-
-    yield
-
-    # Restore original services
-    obsidianki.cli.services.AI = original_ai
-    obsidianki.cli.services.OBSIDIAN = original_obsidian
-    obsidianki.cli.services.ANKI = original_anki
+import tests.utils
+mock_services = tests.utils.mock_services
 
 
 @pytest.fixture
